diff: Don't try to call unlink(NULL) on certain error paths.
Actual bug found by GCC analyzer.

(cherry picked from commit acefa2375868f56fa95e45ddda2df9139cda02d3)

+++ diff/diff.c
@@ -773,8 +773,8 @@ diff (struct file *file1, guestfs_h *g1, struct file *file2, guestfs_h *g2)
   printf ("@@ %s @@\n", _("End of diff"));
 
  out:
-  unlink (tmpda);
-  unlink (tmpdb);
+  if (tmpda) unlink (tmpda);
+  if (tmpdb) unlink (tmpdb);
   rmdir (tmpd);
 }
